リファクタリングと、機能削減によるコードの短縮化

+++ Chapter06/Lesson29/app/src/main/java/com/yokmama/learn10/chapter06/lesson29/net/RequestDownloadImage.java
@@ -2,11 +2,8 @@ package com.yokmama.learn10.chapter06.lesson29.net;
 
 import android.content.Context;
 
-import org.json.JSONException;
-
 import java.io.BufferedInputStream;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -29,8 +26,8 @@ public class RequestDownloadImage {
         return mImageDir;
     }
 
-    public File reqDownloadImageSync(final String urls) throws IOException {
-        final URL url = new URL(urls);
+    public File reqDownloadImageSync(final String url) throws IOException {
+        final URL imageUrl = new URL(url);
 
         // 保存先
         final File imageFile = new File(mImageDir, UUID.randomUUID().toString());
@@ -39,20 +36,17 @@ public class RequestDownloadImage {
                 throw new IOException("ディレクトリの作成に失敗しました。容量エラー？");
             }
         }
-        imageFile.createNewFile();
 
         // HTTP通信
         HttpURLConnection conn = null;
         InputStream stream = null;
         try {
-            conn = (HttpURLConnection) url.openConnection();
+            conn = (HttpURLConnection) imageUrl.openConnection();
             conn.setRequestMethod("GET");
             stream = new BufferedInputStream(conn.getInputStream());
 
-            // ファイルとして保存
-            FileOutputStream output = new FileOutputStream(imageFile);
-
             // 書き込み
+            FileOutputStream output = new FileOutputStream(imageFile);
             byte[] buffer = new byte[1024];
             int len = 0;
             while ((len = stream.read(buffer)) != -1) {
@@ -62,17 +56,19 @@ public class RequestDownloadImage {
             return imageFile;
         } catch (IOException e) {
             // 通信失敗したら、中途半端なファイルを削除
-            imageFile.delete();
-            return null;
-        } finally {
-            if (conn != null) {
-                conn.disconnect();
+            if (imageFile.exists()) {
+                imageFile.delete();
             }
+            throw e;
+        } finally {
             if (stream != null) {
                 try {
                     stream.close();
                 } catch (IOException e) {
                 }
+            }
+            if (conn != null) {
+                conn.disconnect();
             }
         }
     }
